+++ core/server/apps/private-blogging/lib/views/private.hbs
